rename AppPermissionList to PermissionAppList to make it more clear that is the list of apps
