+++ src/index.js
@@ -26,7 +26,7 @@ import AppStore from './js/AppStore';
 import AppStoreMenu from './js/AppStoreMenu';
 import WebEngineAppContainer from './js/WebEngineAppContainer';
 import AppPermissions from './js/AppPermissions';
-import AppPermissionList from './js/AppPermissionList';
+import PermissionAppList from './js/PermissionAppList';
 import Keyboard from './js/Keyboard';
 import React from 'react';
 import ReactDOM from 'react-dom';
@@ -344,7 +344,7 @@ ReactDOM.render((
             <Route path="/appstore" component={AppStore} />
             <Route path="/appstoremenu" component={AppStoreMenu} />
             <Route path="/keyboard" component={Keyboard} />
-            <Route path="/permissions" component={AppPermissionList} />
+            <Route path="/permissionapplist" component={PermissionAppList} />
             <Route path="/apppermissions" component={AppPermissions} />
         </HashRouter>
     </HMIApp>
